Feature Addition 2
Completed first object data string generator method. Also split meta strings into SelfMeta and ContentMeta. SelfMeta describes the element itself so in the case of an object, the SelfMeta would be O"<name>" while the ContentMeta is a description of the insides of an element. This also implies that primitives do not have ContentMeta strings.